requirements.txt
requests>=2.31
fpdf2>=2.7
Pillow>=10
markdown>=3.5
